+++ src/entry/common.ts
@@ -7,6 +7,14 @@ export { ParseError } from '../parse-error';
 export { Unit } from '../units';
 export { ParserEventKind } from '../data';
 
-export type { ParserEvent, FrameFields, MainFrame, SlowFrame, GpsFrame, Stats } from '../data';
+export type {
+	DataParserOptions,
+	ParserEvent,
+	FrameFields,
+	MainFrame,
+	SlowFrame,
+	GpsFrame,
+	Stats,
+} from '../data';
 export type { FrameDef } from '../headers';
 export type { WasmObject, WasmInit } from '../wasm';